refactor(eval): shorten task_args warning; move tests into test_eval.py
Trim the unconsumed-task_args warning to its actionable core, move the
warning tests into the existing tests/test_eval.py (dropping the
standalone test_task_args_warning.py), and note in CLAUDE.md to prefer
adding tests to an existing file over creating a new one.
